Trait tink_testing_server::proto::prf_set_server::PrfSet

pub trait PrfSet: Send + Sync + 'static {
#[must_use]    fn key_ids<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<PrfSetKeyIdsRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<PrfSetKeyIdsResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn compute<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<PrfSetComputeRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<PrfSetComputeResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with PrfSetServer.

Required methods

#[must_use]fn key_ids<'life0, 'async_trait>(
&'life0 self,
request: Request<PrfSetKeyIdsRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<PrfSetKeyIdsResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Returns the key ids and the primary key id in the keyset.

#[must_use]fn compute<'life0, 'async_trait>(
&'life0 self,
request: Request<PrfSetComputeRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<PrfSetComputeResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Computes the output of the PRF with the given key_id in the PrfSet.

Loading content...